requirements.txt
flask
flask-sqlalchemy
python-dotenv
requests
psycopg2-binary
